requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
faiss-cpu==1.13.2
fastapi==0.135.3
ffmpy==1.0.0
filelock==3.25.2
fsspec==2026.3.0
gradio==5.50.0
gradio_client==1.14.0
groovy==0.1.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.4
orjson==3.11.8
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pydantic==2.12.3
pydantic_core==2.41.4
pydub==0.25.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
python-multipart==0.0.24
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==14.3.3
ruff==0.15.10
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
sentence-transformers==5.4.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
tomlkit==0.13.3
torch==2.11.0
torchvision==0.26.0
tqdm==4.67.3
transformers==4.57.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
uvicorn==0.44.0
websockets==15.0.1
